changing my mind (again) on standalone project structure
