+++ README.md
@@ -5,14 +5,9 @@ In early stage of development
 
 ## Building
 
-ResourceManager
+Standalone Cluster Manager
 ```
-$ ./compile.sh resourcemanager
-```
-
-TaskManager
-```
-$ ./compile.sh taskmanager
+$ ./compile.sh standalone
 ```
 
 AppManager
@@ -34,22 +29,22 @@ $ ./compile.sh
 
 Resource Manager
 ```
-$ ./bin/resourcemanager.sh
+$ ./bin/standalone --resourcemanager
 ```
 
 Task Manager
 ```
-$ ./bin/taskmanager.sh
+$ ./bin/standalone --taskmanager
 ```
 
 AppManager
 ```
-$ ./bin/appmanager.sh
+$ ./bin/appmanager
 ```
 
 StateManager
 ```
-$ ./bin/statemanager.sh
+$ ./bin/statemanager
 ```
 ## Testing
