+++ bin/appmanager
