+++ bin/standalone
@@ -1,0 +1,9 @@
+#!/bin/bash
+git config --global alias.root 'rev-parse --show-toplevel'
+REPO_ROOT=$(git root)
+
+[ -f "$REPO_ROOT"/build/standalone.jar ] || { echo "Jar file not found, run compile.sh!" ; exit 1; }
+
+java -jar -Djava.library.path=native/ $REPO_ROOT/build/standalone.jar "$@"
+
+
