--- cluster-manager/standalone/resourcemanager/src/test/scala/clustermanager/standalone/resourcemanager/TestHelpers.scala
@@ -1,8 +1,0 @@
-package clustermanager.standalone.resourcemanager
-
-import runtime.protobuf.messages.{ArcJob, ArcProfile, ArcTask}
-
-trait TestHelpers {
-  val arcProfile = ArcProfile(1.0, 2000)
-  val testArcJob = ArcJob("test", arcProfile, Seq(ArcTask("", "")))
-}
