+++ cluster-manager/standalone/src/main/scala/clustermanager/standalone/taskmanager/utils/Hardware.scala
