+++ cluster-manager/standalone/src/test/scala/clustermanager/standalone/taskmanager/utils/HardwareSpec.scala
