+++ cluster-manager/standalone/src/test/scala/clustermanager/standalone/taskmanager/utils/OperatingSystemSpec.scala
