--- cluster-manager/standalone/taskmanager/src/test/scala/clustermanager/standalone/taskmanager/TestHelpers.scala
@@ -1,9 +1,0 @@
-package clustermanager.standalone.taskmanager
-
-import runtime.protobuf.messages.{ArcJob, ArcProfile, ArcTask}
-
-
-trait TestHelpers {
-  val arcProfile = ArcProfile(1.0, 2000)
-  val testArcJob = ArcJob("test", arcProfile, Seq(ArcTask("", "")))
-}
