+++ project/Sigar.scala
@@ -1,4 +1,3 @@
-
 import sbt._
 import Keys._
